requirements.txt
# Document extraction
pymupdf==1.23.8
python-docx==1.1.0

# Embeddings (free, local)
sentence-transformers==2.3.1
torch==2.1.0

# Vector DB
pinecone==5.0.0

# LLM
groq==0.4.2
requests==2.31.0

# Orchestration
apache-airflow==2.9.3

# UI
streamlit==1.31.0

# Utils
python-dotenv==1.0.0